requirements.txt
Pillow>=8.1.1
pyora==0.3.10
appJar==0.94.0